% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datapkg_read.R
\name{datapackage}
\alias{datapackage}
\alias{datapkg}
\alias{datapkg_read}
\title{Open data-package}
\usage{
datapkg_read(path)
}
\arguments{
\item{path}{file path or URL to the data package directory}
}
\description{
Loads a data and meta-data from a 'data-package' directory or URL.
}
\examples{
# Example data from https://github.com/datasets
datapkg_read("https://raw.githubusercontent.com/datasets/ex-tabular-multiple-resources-fk/master")
datapkg_read("https://raw.githubusercontent.com/datasets/gini-index/master")
datapkg_read("https://raw.githubusercontent.com/datasets/euribor/master")
}
\references{
\url{http://frictionlessdata.io/data-packages}, \url{https://github.com/datasets}
}

