% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_docs.R
\docType{package}
\name{datapkg-package}
\alias{datapkg}
\alias{datapkg-package}
\title{Datasets Galore!}
\description{
Holds various datasets used within the `SMAC-Group`'s packages.
}
\details{
The data sets in this package may change at a moments notice.
}

