% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_region.R
\name{get_region}
\alias{get_region}
\title{Get Region for a City}
\usage{
get_region(city)
}
\arguments{
\item{city}{Character. Name of the city to look up.}
}
\value{
Character. The region name, or NA if city not found.
}
\description{
Looks up the region for a given city using internal package data.
}
\examples{
get_region("Alphaville")
get_region("Gammatown")
}
