% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_docs.R
\docType{data}
\name{hydro}
\alias{hydro}
\title{hydro}
\format{Time series with frequency 12 starting at 1907 and going to 1972 for total of 781 observations.}
\usage{
hydro
}
\description{
Hydrology data that indicates a robust approach may be preferred to a classical approach when estimating time series.
}
\keyword{datasets}

