% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_docs.R
\docType{data}
\name{msft}
\alias{msft}
\title{msft}
\format{Data frame with 5 variables and 127351 observations.
\describe{
\item{Date}{Date of Transaction in \code{MM-DD-YYYY} format.}
\item{Time}{Time of Transation in \code{HH:MM:SS} format.}
\item{Bid}{The bid price made.}
\item{Ask}{The asking price.}
\item{Symbol}{Company symbol.}
}}
\usage{
msft
}
\description{
High frequency MSFT stock data set.
}
\keyword{datasets}

