% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_package.R
\name{write_data_package}
\alias{datapackage}
\alias{read_data_package}
\alias{write_data_package}
\title{Read / write data packages}
\usage{
write_data_package(df, path = ".", format = "csv")

read_data_package(path = ".")
}
\arguments{
\item{df}{a data frame}

\item{which}{format to use for storing data}
}
\description{
Read and write data frames using the standard data packagist format.
}
\examples{
# Write example data
pkgdir <- tempfile()
dir.create(pkgdir)
data(diamonds, package = "ggplot2")
pkg <- write_data_package(diamonds, pkgdir)

# What it looks like
list.files(pkgdir)
cat(readLines(pkg), sep = "\\n")

# Read it back
#mydata <- read_data_package(pkgdir)
#all.equal(ggplot2::diamonds, mydata)
}

