% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_package.R
\name{data_package}
\alias{data_package}
\alias{datapackage}
\title{Data-package}
\usage{
data_package(path = ".", verbose = TRUE)
}
\arguments{
\item{path}{root directory of the data package}

\item{verbose}{emits some debugging messages}
}
\description{
Load or initiate a \href{http://dataprotocols.org/data-packages}{data package} for
reading / writing data and metadata. A data package can be an R package at the same
time. The default format for storing data is
\href{http://dataprotocols.org/linear-tsv}{linear-tsv} which is the least
ambiguous format and natively supported by R via \code{\link{read.table}}
or \code{\link[readr:read_tsv]{readr::read_tsv}}.
}
\examples{
# Create a data package in a dir
pkgdir <- tempfile()
dir.create(pkgdir)
pkg <- data_package(pkgdir)

# Show methods
print(pkg)

# Examples
pkg$author("Jerry", "jerry@gmail.com")
pkg$resources$add(iris)
pkg$sources$add("Fisher, R. A. (1936)")

# View json file
pkg$json()

# Parse data
pkg$resources$read("iris")
}

