% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datapkg_read.R, R/datapkg_write.R
\name{datapackage}
\alias{datapackage}
\alias{datapkg}
\alias{datapkg_read}
\alias{datapkg_write}
\title{Open data-package}
\usage{
datapkg_read(path)

datapkg_write(data, name, path = getwd())
}
\arguments{
\item{path}{file path or URL to the data package directory}
}
\description{
Retrieve and parse data and meta-data from a 'data-package' directory
or URL.
}
\examples{
# Clone package with git:
cities <- datapkg_read("git://github.com/datasets/world-cities")

# Read over http
euribor <- datapkg_read("https://raw.githubusercontent.com/datasets/euribor/master")
}
\references{
\url{http://frictionlessdata.io/data-packages}, \url{https://github.com/datasets}
}

