% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datapkg_read.R, R/datapkg_write.R
\name{datapackage}
\alias{datapackage}
\alias{datapkg}
\alias{datapkg_read}
\alias{datapkg_write}
\title{Read/write data-package}
\usage{
datapkg_read(path = getwd())

datapkg_write(data, name, path = getwd())
}
\arguments{
\item{path}{file path or URL to the data package directory}
}
\description{
Read and write data frames to/from 'data-package' format. For reading
supported paths are disk, http or git. For writing only disk is supported.
}
\examples{
# Create new data package
pkgdir <- tempfile()
datapkg_write(mtcars, path = pkgdir)
datapkg_write(iris, path = pkgdir)

# Read it back
mypkg <- datapkg_read(pkgdir)
print(mypkg$data$mtcars)

# Clone package with git:
cities <- datapkg_read("git://github.com/datasets/world-cities")

# Read over http
euribor <- datapkg_read("https://raw.githubusercontent.com/datasets/euribor/master")
}
\references{
\url{http://frictionlessdata.io/data-packages}, \url{https://github.com/datasets}
}

